renamed and updated test about the lazy CSRF
